requirements.txt
llama-cpp-python
gradio
transformers
peft
datasets
bitsandbytes
torch
sentencepiece